id #10157
There’s a new functionality here.

Users will be able to inject css and js on mantle[PUC]. To do so,
they’ll need to create into their template folder, a json file named
template-config.json and add to it an array called “externalResources”
of elements of the type:
{
 type : [String | “css”,”js”]
 context : [String | “mantle” ] this can be used later for more contexts
 url : [String | the url relative to the template folder]
}

Unfortunately, when these resources are loaded, there’s only jquery
1.4.4 available, so the 404 handling is… nonexistent.

I’ve added to tapa-default template, an example that places a logotype
at pucHeader.