Big push
Big push to Tapa. There’s a whole new Admin-interface to make user’s
life easier.

With this interface, users are able to:
 - enable/disable splash-screen for the admin-ui - which is not yet
implemented, though controls are available;
 - enable/disable plugin - this switches PUCLogin.jsp file;
 - Choose to enable available templates - interface shows available and
marks the one which is active;

It also delivers through assets injection on some templates, a fix to
PUC described in http://jira.pentaho.com/browse/BISERVER-12628;

I’ve put this admin-ui into an admin-perspective to get more
visible-height on the viewport.