Fixed issue when renaming and changing pictures at the same edit session.
When editing, when you changed the template name and then tried to
upload a new picture, the picture was being copied to the “new folder”
that still didn’t exist.

So instead of using ctrl.current, which is binded by the rename input,
we used the _original template object that is kept on the beginning of
the edit session.